+++ EngineSource/Rendering/Camera/CameraShake.cpp
@@ -1,5 +1,6 @@
 #include "CameraShake.h"
-#include <World/Stats.h>
+#include <Engine/Stats.h>
+
 namespace CameraShake
 {
 	Vector3 CameraShakeTranslation;
@@ -20,8 +21,8 @@ void CameraShake::StopAllCameraShake()
 void CameraShake::Tick()
 {
 	float ModifiedShakeIntensity = powf(ShakeIntensity, 2);
-	CameraShakeTranslation = Vector3(sin(Stats::Time * 30) * ModifiedShakeIntensity, sin(Stats::Time * 25) * ModifiedShakeIntensity, 0);
-	ShakeIntensity = std::max(ShakeIntensity - Performance::DeltaTime * 5, 0.f);
+	CameraShakeTranslation = Vector3(sinf(Stats::Time * 30) * ModifiedShakeIntensity, sinf(Stats::Time * 25) * ModifiedShakeIntensity, 0);
+	ShakeIntensity = std::max(ShakeIntensity - Performance::DeltaTime * 5, 0.0f);
 }
 
 float CameraShake::GetCurrentCameraShakeIntensity()
